增加删除空员工的逻辑判断

+++ src/main/java/com/zzs/service/impl/EmpServceImpl.java
@@ -65,8 +65,9 @@ public class EmpServceImpl implements EmpService {
     @Transactional(rollbackFor = {Exception.class})
     @Override
     public void delete(List<Integer> ids) {
-        empMapper.deleteByids(ids);
-        empExprMapper.deleteByEmpids(ids);
-
+        if(!CollectionUtils.isEmpty(ids)){
+            empMapper.deleteByids(ids);
+            empExprMapper.deleteByEmpids(ids);
+        }
     }
 }
